- schema_creation class got a new file for autoload - readme updated
